Dreaded balance guy towner nerfing (#2039)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request


One of the main arguments for why towners had to have weapon skills on
par with adventurers was that they were very prevalent and strong; now
that refugees exist, they are all weaker than the adventurers of old,
thus the need for towners to possess strong weapon skills has
diminished. Towners should not randomly have skills on par with people
who are trained to fight.

## Why It's Good For The Game

Powercreep is a real thing.