Adjusts Druid & Mortician Stats/Skills to Better Reflect Their Roles (#1981)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Nerfed mortician's close quarters capability a tiny bit, they get better
treatment skills too.

Buffed druid farming and made them in-line with the rest of clergy. I
reverted a lot of changes I made before since it was received poorly.

## Why It's Good For The Game

Druids actively do farm work or running about, so it makes sense for
them to do it slightly faster. Buffed their combat to be on par with the
rest of the church team.

Morticians should be able to put someone's head back on if needed should
they get the tools. Previously they had better combat skills than
acolytes which was crazy so that got nerfed by one level to be as good
as the rest of the church team.

**Please give me input if this needs changing in any way.**