Removal of Player Occultists
Skeletons are a broken mess and unfun to fight against.
I absolutely adore occultist in idea and practice, but skeletons themselves are a mess and need to be cleaned up. As a result, this PR decouples it from the magos refugee altogether and makes it an outfit for admin intervention exclusively.
That's about it.