Occultist / Zizoid Mage Rebalance (#2054)
## About The Pull Request
Skeletons are a broken mess, both unfair and unfun to fight against.
I absolutely adore occultist in concept and practice, but their minions
break the subclass entirely and thrust it into the realm of genuinely
cancerous in nature.
As a result, this PR decouples old Occultist from the magos refugee
altogether and makes it an outfit for admin intervention exclusively.

Now, occultist in its state for this PR, admin exclusive outfit aside,
has no such abilities to raise the dead. It can still heal them, and
possess the strengthen undead stun spell, but it can't raise corpses
anymore.

'Raise Undead', 'Command Undead' and 'Revoke Unlife' are no longer
purchasable for Zizoid mages.

'Strengthen Undead' cost raised from 2 to 4. It was far too powerful
utility and combat wise to be at TWO.
'Eyebite' is now purchasable for Zizoid mages with a cost of 3, because
it was missing. This makes 'Ray of Sickness', 'Eyebite' and 'Strengthen
Undead' the spells added to the list of possible to purchase for a
Zizoid mage.

We'll leave the hoards of unkillable creatures to Lich, outside of admin
curated events / intervention.

## Why It's Good For The Game
A myriad of player and administrative complaints have prompted me to
push this. As someone who plays both skeleton AND fights them regularly,
I'd like to think I've a good idea of both sides, and this is the best
solution at the moment.
Barring a full skeleton rework, at least, but the Ten knows I refuse to
touch that with a ten foot pole.