Nerve Master Withdrawing (#2047)
## About The Pull Request

Adds a simple button to the Nerve Master to let anyone withdraw money
directly from the treasury.


![image](https://github.com/user-attachments/assets/f3e7fafb-c985-442b-9e1e-b2e44255fa49)

## Why It's Good For The Game

When bandits stopped being able to create Shylock accounts, they stopped
being able to rob the treasury themselves. This simple feature allows
them to do this optional objective once again.

This also stops the pointless task of stewards needing coins in hand and
paying into their own accounts, and then running to a Shylock to
withdraw the money that way, which happens a lot.