More Fisherman Changes (#2065)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This time, it's a slightly larger re-buff to make the loot table less
frustrating but still random enough.

More chance of rings again, less chances of coin pouches; less chances
of useless junk like clothing, shards, etc
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People thought the nerfs were too harsh, I'm inclined to agree but not
for the same reasons. Removing the double-stacked chances of rings on
worms seemingly almost entirely fixed the rareness (or lack thereof)
issue and this ended up being unusually harsh due to the high pick
weights (relatively) of other items.

At some point this'll be reworked entirely into a full fishing system
instead of just a gacha table, but for now, this should be fine.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->