Fixes the leather coat crafting recipe (#2070)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I may or may not have forgotten to add a little "/heavy" and the proper
skill requirement at the end of the crafting recipe when i made the
original PR, so this fixes it!

Oh, and this also lists the materials needed for the crafting recipe on
the name.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Bug fix!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->