Makes ecclesiast accessible to shunned races (#2078)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

So far, every single refugee role is allowed to all races, but
ecclesiast, that for some reason has even more race restrictions than
the local church of the ten, which lets templars and acolytes of any
races be a part of it, other than the priest, being the only role at the
church that needs to be at least tolerated.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More ecclesiasts around meaning more diversity on the refugees, and
still getting no respect or honor whatsoever for being a foreigner and a
shunned race too.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->