Returns Necromancy to Necromancers (#2073)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This pull request returns the less raise undead spell to the Occultist
refugee. This change comes with a nerf to the raised skeleton, as well
as restrictions, and a small buff to the ray of sickness spell. It also
gives skeletons non-combat skills so they can act in other roles besides
combat, hopefully pushing the class away from just a frag role.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Occultist right now, with raise dead disabled, is basically useless. It
has no spell points to learn spells, one of its spells is useless
because it has no more undead to raise, and ray of sickness as far as
I've tested and been told by those who use it is pretty underwhelming
for a 3 point spell.

This PR returns the ability to raise skeletons to the Occultist while
also making it impossible to raise bums, as those were the mobs that
players had the most issues with once skelliefied. Some outright lost
crit weakness and gained crit resistance, making them essentially
immortal. It also returns the ability to revoke undeath as suggested, so
to stop shitter skeletons on their tracks or un-do minions that are no
longer useful.

The PR also gives all raised skeletons crit weakness, meaning that
critical strikes will cripple or outright instantly kill any skeletons
hit by them.

I've tested it in my own local server, everything I raised was able to
be easy killed and beheaded. Bums, even rotten and deadified, are no
longer able to be raised.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->