Court physician patron lock removal (#2096)
Removes court physician patron lock. So you may choose a Patron besides
Pestra or Graggar. Also remove the Pestra followers "Rot eater" trait
from court physician as it was force added as a trait even if you
followed Graggar

From my knowledge of the lore, this Patron lock was in place due to the
old tutorial which at some point was change.

This tutorial mentioned you were a sickly child, stricken with leprosy,
so you prayed and either Pestra or Graggar removed your leprosy so you
started following in Pestra or Graggar, learning physician work

The below quote is what the tutorial is now. Mentions you following the
teaching of Pestra, not you following Pestra herself. This aligns it to
match towner physician allowing the follow of all Patrons even though
you are following Pestra's teaching

> You were a child born into good wealth - But poor health. Perhaps in
another life, you would have turned out to be a powerful mage, wise
archivist or a shrewd steward, but leprosy took away your younger years.
Out of desperation, you followed the ways of Pestra and managed to be
cured. Now you serve in the Duke's court ensuring the good health of
those inhabiting the keep.


![image_2025-01-01_124859393](https://github.com/user-attachments/assets/a5a7acd9-42d4-487c-a2d6-1269187c1ebd)