Arrow/Bolt Crafting Tweak & Hunter/Tracker Skill Adjust (#2093)
## About The Pull Request

Gives Smiths arrow/bolt crafting back and makes arrows and bolts be
crafted in stacks of 20 for 2 ingots. Swordfighting is removed from
Hunter/Tracker and replaced with Polearms skill. Crafting skill is
increased to 2 for Hunter and 3 for Tracker.

## Why It's Good For The Game

This makes arrow/bolt crafting not a pain in the bum in terms of time
and resource inefficiency, it also means that Hunters can refill on
arrows a little cheaper and without needing an Artificer present - a
Smith can do just fine. Since they don't actually _start_ with a hunting
saber, and neither is one present in the code, swordfighting is instead
replaced with polearms - spears are a fitting alternative way of hunting
animals. Crafting skill is also bumped up so that Tracker can craft a
Stone Spear for himself, and a Hunter only needs to level up Crafting
once to do the same, if there are no Smiths or Artificers present.

Might edit stone arrows to also be similarly bulk-craftable, as at
present they are so much of a pain to make as to not be worth it.