Hoardmaster Salt-PR Gutting
Time for the bi-weekly anti-bandit salt PR. Praise be!

Effectively removes 90% of Hoardmaster items for classes.
Strips it down to the bare minimum, so it's things they NEED to operate, at best, and a few sidegrades / upgrades here and there.
Bandits will now once again have to rely on loot, as opposed to favour farming from the zizoid church smelt strat, the well-deservedly nerfed fishing abuse, and the myriad other methods to gain favour that I really don't want to repeat here, otherwise it'll be used for farming, too.

We really ought to make it so you can't open other player's menus, as a lot of these issues would be solved by that alone, and I wouldn't need to make a PR like this. I stand by the Hoardmaster being a mistake.

As an aside, this also gives them a light gambeson option to purchase, cheap as dirt, as a consolation prize.

I originally had this paired with another change, that being the inclusion of the Forgemaster, within #2023. That wasn't wanted because of the Forgemaster and so on, as can be see in the eventual PR when I decoupled it. As a result, this PR is effectively the same, and hasn't taken into account any further changes from the hoardmaster rebalances (if any) that have taken part since.

Keep in mind, this doesn't touch ANYTHING regarding Stuff / Things as categories. They're whatever they were prior, given they weren't touched in the interim of the initial PR and now this, the exception being the included gambeson.