Steward glasses removal + no longer forced to be virgin (#2113)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The steward is no longer a virgin and is forced to be blind.
We are a roleplay server, why the fuck are all the stewards forced to
have glasses and be virgins.
I can remove the +1 to speed if needed. But, the steward is a role that
does a lot of running, it makes sense that he at least has one speed.

Edit: I added the glasses back so that they can be worn as a choice,
instead.

![image](https://github.com/user-attachments/assets/c02c1658-c65c-494a-8d45-a308e9254800)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having forced traits is horrible, and having to be a virgin and
nearsighted for the "haha, funny" is childish coding at best.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->