Removes the ability to use triumphs to resist confession. (#2137)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removes the prompt for people to pay one triumph to resist confession.
Replacing it with a percentage chance to resist. Antags get an
additional 25% to resist.

Formula is...

INT + CON + 10 (+25 if antag)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

A mechanic that derived entirely around getting confessions with RP and
mechanics leading up to it should not be thwarted by something can be
accrued across dozens if not hundreds rounds. One triumph isn't even a
ripple in the proverbial pond.

No flaw? Makes sense, 3 triumphs.
Mood boost? Weird but eh... alright?

Resisting Confession? No, why? This doesn't make sense this is a
mechanic that requires another player (one who is irreplaceable) and has
to even find the confessed(to be) to interrogate in the first place.
Using something that is accrued over multiple rounds.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->