The Glorious Evolution (#2097)
## About The Pull Request
Allows you to step closer to discarding your weak and rotting flesh to
achieve mechanical perfection. From the moment I understood the weakness
of my flesh, it disgusted me.

Firstly, buffs all prosthetic limb durabilities to match a person of
average or above average constitution.

List of all limbs
Wood:
Flammable, has 100 health and cannot move the fingers, the cheapest of
the bunch.

Bronze:
Has 170 health and can move the fingers, an objective upgrade from iron
and gold but can only have arms.

Gold:
120 health and is a downgrade to bronze, less health but can have gold
legs.

Iron:
150 hp, as bulky and clunky as wood but a tad harder to source requiring
an artificer. Comes with a 5 damage reduction on brute and burn.

Steel:
The most durable prosthetics at 200 hp, the best for tanking but has no
fingers similar to wood and iron. Comes with a 10 damage reduction on
brute and burn.

Limbs cannot be healed via potions or sleep but require miracles,
surgery or a smith to repair.

Prosthetic limbs cannot bleed.

Added a windup to removing limbs.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More use for the artificer crafts, current limb healths are too low to
be viable given that flesh limbs can be transplanted easily by any
surgeon or acolyte.
Also adds a way to repair limbs as well as removing an exploit that
makes you able to pop off a dudes prosthetics if you get a solid grip.
Enables repairing limbs for QoL.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->