Round wont start without a Duke/Duchess (#2132)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Prevents round from starting if nobody has readied up as a Duke/Duchess.
This is ignored if FASTLOAD is defined or if an admin overrides it with
the Start Now button.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Prevents round start bad omens and generally makes the round better by
always having a Duke.