Adds A New Lighter Color to Horc Clan Skintones (#2157)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Added a nice, fitting green color to the horc skin tones called 'light
raider'. I couldn't think of any better name to be honest, but it
sounded cool and indicates its a lighter tone while keeping the rough
sounding names of clans.

## Why It's Good For The Game

All of the skin tones for horcs are shit and too dark. The only light
skin one we have is red and it would be nice to have one that isn't
extremely dark green or blackish since we currently can't change skin
tones like you can with wild kin on horcs.

Tested it locally, looked good.


![tone2](https://github.com/user-attachments/assets/1a653d38-e55d-4b1a-90b1-c07cc1ae01d5)