nerfs wooden platform cost, makes carpentry and masonry easier  (#2149)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

removes fiber costs for wooden platforms and stone

## Why It's Good For The Game

the requirement for fibers on wood and stone made it so that almost
every time someone built would build twig platforms first instead of
bothering to gather the fibers. atop of that it just makes more sense to
need fibers to tie twigs together

as a second twig platforms are so accessible that often times people
will make long twig bridges to get past walls instead of getting a
proper carpenter or mason

# Changes
raised stone platform cost to 2 stone but felt like it wasnt enough, so
additionally raised the skill level to stone and wooden platforms to
level 2 (same as walls)