Leather armor & newmoon nerf (#2139)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Lowers the durability of most leather armor to make more sense,
considering some exceeded that of plate and other heavy armors. Heavily
nerfs the newmoon coat so it's not outright broken, making it similar to
the hardened leather coat.

- Reinforced leather hood:
  - Durability: 200 -> 150
  - CRITS: blunt, twist -> cut, blunt, twist
- Hardened leather coat:
  - Durability: 350 -> 200
- Masterwork leather coat:
  - Durability: 400 -> 300
- Hide armor:
  - Durability: 200 -> 175
- Newmoon Jacket:
  - Durability: 300 -> 225
  - CRITS: cut, stab, blunt, chop -> cut, blunt, twist
- ARMOR: "blunt" = 70, "slash" = 100, "stab" = 60, "bullet" = 15 ->
"blunt" = 70, "slash" = 60, "stab" = 30, "bullet" = 10