Separate the Inquisition into its own category. (#2153)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Title.


![image](https://github.com/user-attachments/assets/2117bcb6-8916-4885-89a1-b02f740a43e3)

![image](https://github.com/user-attachments/assets/e7a9098f-14a6-4c41-9b4e-3b2540926f94)

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

A style change to better reflect the reality of things, the Inquisition
isn't part of the Church, despite working with it.