Vampire Lord rebalance (#2147)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

VL sword (Crimson fang):
- Force: 40 -> 30
- Force Wielded: 55 -> 38

Vampire:
- Shapeshift into gas and bat now have a charge time of 5 seconds.
- Axed regenerate and replaced it with vamp rejuvenate
- Rejuvenate added as a spell
- Bat / Gas form lowered from 2000 -> 750
- Swift intent removed
- Vitae doesn't drain passively unless you are disguised.
- Being disguised makes you not be affected by miracles (no more spam
casting it to check if someone is a vampire...)

Deathknight / Skeleton (players):
- Given more intents (not including swift)

Vamp Rejuvenate:
- 250 cost
- recharge time is 60 seconds
- Announces a visible message of the user being shrouded in shadows.
- Heals wounds by the heal formula... (below)
- Heal formula Lesser: (roll("[bloodskill]d8") + (vampire.STACON * 1.5))
* 2
- Heal formula lord: (roll("[bloodskill]d10") + (vampire.STACON * 2)) *
3


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Regenerate was axed because ahealing in combat with no cooldown and only
a vitae cost was overpowered. It being replaced with something free and
on charge but heals WAY less and is based on blood skill seems to fit
more in line with the rest of the balancing we have right now.

Vampire lord sword was reduced in force due to the fact it had such a
high decap chance; tied in with the infinite swift intent attacks made
for a nasty combination.

After discussion with multiple people; the issue didn't lie in the
infinite stamina but rather the swift intent itself. (1/7/2025)