Racelocked Tabaxi noises without 30 useless commits (#2177)
## About The Pull Request

Same as the PR I closed due to audio quality concerns. Adds meows,
trills, hisses and purrs for Tabaxi only, with two variations for each.

Half-kin aren't getting this. They will be receiving their own sets of
audio if I ever decide to PR it. Wild-kin may, possibly, just maybe get
this in the future.

## Why It's Good For The Game

Noise can be good. Spamming noise is NOT so good. Please TM this to see
whether or not the Tabaxi playerbase can be trusted with furry noises.