Adjusts Bard Hat / Fancy Hat Recipe Slightly + Gives Cows Fat (#2195)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Made the bard hat require less skill at three but require a feather and
instead of cloth it is leather now. Kept fancy hat the same but added
feather to the recipe as well. I also gave cows the ability to drop fat
when butchered because they were the only animal that didn't and that
was weird.

## Why It's Good For The Game

Bard hat looks more like a leather hat and it isn't as fancy as the
fancy hat so three seems fine. Fancy hat is fancier so it stays at four.
Also giving the recipes a feather which they both have in their sprites
allows those who want to make them to ask the soilson or butcher to get
them a feather which is a very underused item outside of writing with.

Also cows didn't drop fat when butchered and that is weird because cows
are fat... and meaty. Chickens dropped a single fat when butchered which
are way less fattier and so did goats so I felt that giving cows some
fat makes sense honestly.