Small fixes to energy recovery chat spam and unable to sleep causes (#2203)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

… sleep fixes, energy recovery spam fixes. Gives both a 30 second delay.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Fixes an unintended behavior.

![image](https://github.com/user-attachments/assets/41d76213-f8fa-4d33-b7dc-c75d346de637)


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->