Moves Physician and Magician under noble like their jobs state they are (#2176)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/def9109f-718c-4a52-9238-4b52f7352653)

![image](https://github.com/user-attachments/assets/27dc44c4-0185-437b-9ed0-49e7bb5433c6)
both physician and magician possess flags that state they are noblemen
the backstory for the physician also them having a gem ring as an
unmarried person fits for a noble, the physician also used to be under
noble category in prefs
and still is a noble under latejoin

this PR aims to properly move the phycisian and magician under nobles by
changing the file pathing and giving them the proper traits to fully
show as a noble

## Notes
Old magician no longer gets arcane casting 1 and now gets a golden belt
to show their value (old age only)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The magician and physician are both flagged as nobles, yet they are both
tagged as a noble and as a courtier in multiple diffrent places
They are now nobles, this also clears up most code making people
confused whetever they are nobles or not

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->