Quick bandaid Fix For Leash Associated Bug (#2215)
## About The Pull Request

This is a quick bandaid fix for the fact that someone broke non-storage
objects being in hipslot somehow. I intend to go over and do a proper
refactor later because it is definitely not intended behavior for
hipslot to only work with /storage objects. All this bandaid does
however is prevent sticking leashes in the hipslot.

## Why It's Good For The Game

Bugfix good.