Mercenary weapon choice (#2207)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Removes the nasty RNG elements when appearing for a mercenary and allows
players to choose something specific for their class.

## Why It's Good For The Game

RNG is a nuisance. It's better to let players choose something they
think is better suited for the character than to pray to Armok in hopes
that you'll get what you need.