Soaps, Sewers, and Soothers
Fixes soap so that all variants are able to be used for bathing. Provides a small buff for bathing in general. Provides a larger buff for bathing by 'specialists'.

Also gives you a negative moodlet for stomping through sewer water, because that makes sense.

Fixes a typo in the stressevent info tab, where the values were flipped