Having no Duke no longer spam pings (#2226)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Sets the current state of the game back to PREGAME rather than STARTUP,
which would cause the bot to ping people.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game


![image](https://github.com/user-attachments/assets/a3c46ac2-95b3-4461-bbcb-63e3d7dafdd3)


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->