Disables Maniac (#2220)
## About The Pull Request
Huzzah. We're free from the mockery of this bargain brand Dreamer.

Simply comments it out, in case we want to return it in the future, for
whatever reason.
I considered removing the code relating to latespawns as well, but that
is already set to return before it does anything, so it's not even
something to bother with..

## Why It's Good For The Game
Do I really need to explain this?