Leather coats are now cheaper (#2221)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

While i was making the leather coats, i made them expensive to make, a
bit... too expensive actually.
The normal leather coat needed so much cured hide, that most seamsters
didn't even know it existed since it never showed up on the crafting
menus, not only that, but it had the highest amount of needed hides out
of all the other things.
And the heavy variant needed 4 furs, which i did not realise made it the
same or even more expensive than FULL STEEL PLATE ARMOR when it was time
to go ahead and buy the materials.
So yeah, hopefully this'll make more of an option, instead of a fancy
way of throwing your mammon away and confusing the seamster after they
make a cloak thinking you just mispelled it when making an order.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More options for the light armor enjoyers that aren't ridiculously
expensive or obscure now, and also more variety on what clothing people
wear.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->