Adds Lynx ears & tail to Tabaxi (as well as Half-Kin and Wild-Kin) (#2230)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The name says it. I was bored, I realized that Tabaxi's lore states that
**like most felines**, they have short tails. Ergo, why not add short
tails? In fact, why not add a short tail that's pointing down for a
change?

The Lynx Ears are a bastardization of Cat Big ears. The tail was made
from scratch.

I have 0 spriting knowledge and barely a clue code-wise, but I did test
it on my end and even made sure it looks *somewhat* good for different
'types' of characters.

Like warm colors and male bodies

![image](https://github.com/user-attachments/assets/3d79e656-2ded-4c93-8261-b096381a8083)

Male bodies and relatively-unusual-colors

![image](https://github.com/user-attachments/assets/14275b22-6758-4ff0-899f-df01c74cbd4f)

And even female bodies with cold colors

![image](https://github.com/user-attachments/assets/9781e260-f02b-49e5-9290-fe34ad215a10)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More customization not restricted to just Tabaxis.

Yes, this works for wild-kins and half-kins too.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->