Underdark Return
Returns my beloved underdark, and all the horrors that accompany it.
Moles, skallax spiders, shrooms, etc.

It was gutted to shave off a few seconds of startup time. Something that doesn't really matter when we've a ten minute lobby timer as is.
In any case, I cleaned up the code a tad (replaced spaces with indents) and assured it functions in game. Reintroduced an entrance at the lava mines into the underdark, alongside removing all surface shrooms. Didn't want to touch too much else, for fear that it'd be difficult to have others copy-paste their changes atop this, if it's merged before their changes.

The moles yearn for the lava mines.