Grenzelhoft Lore update + Grenzelhoft Jager + Firearms Nerf + Firearm File Cleanup (#2154)
## About The Pull Request

Changes a lot of descriptions for the Merc Grenz stuff.

Adds a Flintlock, essentially the inbetween of a Handgonne and a
Arquebus.
Adds Grenzelhoft Jager role. It's not a proper Frontline Melee like the
others, meant to be filled by a lot of random races.

![image](https://github.com/user-attachments/assets/02317f45-d94c-4e9c-b6c4-ab571f72534b)

Nerfs Firearm aim time across the board. 

CLEANS FIREARM CODE

## Why It's Good For The Game

Lore + Adds requested Firearm Merc.

No more hitscan, instashot firearm that doesn't require to be aimed.
Horrible to play and fight against.
It feels a bit WEIRD that you can readily aim a firearm without sights
this accurately.