The glorious fixes (#2243)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes some bugs regarding repairing prosthetics, removes the ability for
prosthetics to get any wounds such as bone fractures or dislocations

During testing prosthetics can be dismembered and disabled with enough
damage

Runtimes were created by all of the wound healing options requiring the
limb to be attached to a mob and I couldn't figure out how to make a new
proc to remove all the wounds. So I simply made it so that prosthetic
limbs can't get wounds.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
General bugfixes and runtimes
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->